cmd/vfsgendev: Simplify code layout further.
Rename run to goRun, use more common run for main command body with
error return value.

Use main.go for file with func main().

Move docpackage.go code into parse.go, since that's where it's used.
