+++ cmd/vfsgendev/generate.go
@@ -1,11 +1,6 @@
 package main
 
 import (
-	"fmt"
-	"io/ioutil"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"strconv"
 	"text/template"
 )
@@ -42,31 +37,3 @@ func main() {
 	}
 }
 `))
-
-// run runs Go code src with build tags.
-func run(src string, tags string) error {
-	// Create a temp folder.
-	tempDir, err := ioutil.TempDir("", "vfsgendev_")
-	if err != nil {
-		return err
-	}
-	defer func() {
-		err := os.RemoveAll(tempDir)
-		if err != nil {
-			fmt.Fprintln(os.Stderr, "warning: error removing temp dir:", err)
-		}
-	}()
-
-	// Write the source code file.
-	tempFile := filepath.Join(tempDir, "generate.go")
-	err = ioutil.WriteFile(tempFile, []byte(src), 0600)
-	if err != nil {
-		return err
-	}
-
-	// Compile and run the program.
-	cmd := exec.Command("go", "run", "-tags="+tags, tempFile)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	return cmd.Run()
-}
